added firebase helper class and code reusability
#added firebase helper class to retrieve all the categories
#used the created helper class in addincomeexpenseactivity and bucks activity
#added category retrieval and totals code with interfaces
#code is reused and a lot of code is reduced and optimized

+++ app/src/main/java/com/chaitupenjudcoder/datapojos/CategoriesAmount.java
@@ -4,12 +4,12 @@ public class CategoriesAmount {
 
     private String categoryName;
     private String totalAmount;
-    private int percentage;
+    private float percentage;
 
     public CategoriesAmount() {
     }
 
-    public CategoriesAmount(String categoryName, String totalAmount, int percentage) {
+    public CategoriesAmount(String categoryName, String totalAmount, float percentage) {
         this.categoryName = categoryName;
         this.totalAmount = totalAmount;
         this.percentage = percentage;
@@ -23,7 +23,7 @@ public class CategoriesAmount {
         return totalAmount;
     }
 
-    public int getPercentage() {
+    public float getPercentage() {
         return percentage;
     }
 }
